Mika v3: drm/i915: Avoid tweaking evaluation thresholds on Baytrail v3
Please test this patch.

drm/i915: Avoid tweaking evaluation thresholds on Baytrail v3

This patch is expected to have the same function as the previous version, which was attached to comment #683.

It should apply, with offset, from Linux-4.2 through Linux-4.10, and now 4.11-merge/rc.

If we can supply some "Tested-by: " tags, that may help Mika get permission to ship this patch with the i915 tree.